updated erizoAPI compilation for cleanup

+++ erizoAPI/build.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
-rm -rf build
-node-waf configure build
+if hash node-waf 2>/dev/null; then
+  echo 'building with node-waf'
+  rm -rf build
+  node-waf configure build
+else
+  echo 'building with node-gyp'
+  node-gyp rebuild
+fi
